pyproject.toml
[project]
name = "rpg-player"
version = "0.1.0"
description = "A small toy project for playing & narrating RPG-style messages"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Lyndon Armitage" }]
dependencies = [
  "elevenlabs>=2.15.0",
  "jinja2>=3.1.6",
  "openai>=1.99.9",
  "piper-tts>=1.3.0",
  "python-dotenv>=1.1.1",
  "pyttsx3>=2.99",
  "sounddevice>=0.5.2",
  "soundfile>=0.13.1",
  "textual>=5.3.0",
  "tiktoken>=0.11.0",
]

[dependency-groups]
dev = [
  "black>=25.1.0",
  "flake8-bugbear>=24.12.12",
  "flake8>=7.3.0",
  "isort>=5.12.0",
  "pre-commit>=4.3.0",
  "pytest>=8.4.1",
  "ruff>=0.12.0",
  "textual-dev>=1.7.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "C", "N"]
exclude = ["src/rpg_player/__pycache__"]

[tool.setuptools.packages.find]
where = ["src"]
